Makes another attempt at fixing wp-load loading on Windows Server.
This addresses the possibility that wp-load fails to load on Windows Server because of the final forward slash. I'm not 100% confident that this will fix the issue, becauase my understanding is that Windows is usually agnostic to forward slashes and back slashes in paths.

Also updates the version number in the plugin metadata.