Re-organizes for subversion-hosted WordPress plugin directory.
This re-organizes the code according to the standards of the plugin directory, i.e. four top-level folders including "trunk" for development code and version-specific folders in "tags".

Subversion version control is also required by the directory, so there is now a .svn folder that lives alongside the existing .git folder, and contains a separate set of commits. I'm not sure what the appropriate workflow is for maintaining code in two parallel version control systems.

This also adds a readme.txt file, which follows the formatting required to generate the plugin page at https://wordpress.org/plugins/wp-placespeak-connect/.

The README.md file is retained so that there will be some explanatory text in the GitHub directory, but it now copies the readme.txt content, which is almost-but-not-quite markdown, so it will look a little weird.